requirements.txt
numpy>=1.9.1
scipy>=0.15.1
PySoundFile>=0.7.0
matplotlib>=1.4.2
PyYAML>=3.11
